fix #1792

+++ src/Snap.Hutao/Snap.Hutao/Service/Game/Configuration/GameConfigurationFileService.cs
@@ -26,8 +26,15 @@ internal sealed partial class GameConfigurationFileService : IGameConfigurationF
     {
         string serverCacheFolder = runtimeOptions.GetDataFolderServerCacheFolder();
         string source = Path.Combine(serverCacheFolder, ConfigurationFileName);
+
         if (File.Exists(source))
         {
+            string? directory = Path.GetDirectoryName(destination);
+            if (string.IsNullOrEmpty(directory) || !Directory.Exists(directory))
+            {
+                return;
+            }
+
             File.Copy(source, destination, true);
         }
     }
